Renaming the 'rest' package name
