+++ aftership-sdk/src/main/java/com/aftership/sdk/AfterShip.java
@@ -13,8 +13,8 @@ import com.aftership.sdk.error.ErrorType;
 import com.aftership.sdk.exception.ConstructorException;
 import com.aftership.sdk.model.AftershipOption;
 import com.aftership.sdk.model.RateLimit;
-import com.aftership.sdk.rest.ApiRequest;
-import com.aftership.sdk.rest.ApiRequestImpl;
+import com.aftership.sdk.request.ApiRequest;
+import com.aftership.sdk.request.ApiRequestImpl;
 import com.aftership.sdk.utils.StrUtils;
 import lombok.Getter;
 import lombok.Setter;
