+++ aftership-sdk/src/main/java/com/aftership/sdk/request/DataEntity.java
@@ -1,4 +1,4 @@
-package com.aftership.sdk.rest;
+package com.aftership.sdk.request;
 
 import com.aftership.sdk.error.AftershipError;
