+++ aftership-sdk/src/main/java/com/aftership/sdk/request/RequestConfig.java
@@ -1,4 +1,4 @@
-package com.aftership.sdk.rest;
+package com.aftership.sdk.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
